Pull request #3465: [BLUETOOTH-15107][KW45] Check and fix MISRA for all digital_key and localization projects
Merge in CONNINT/bluetooth_public from BLUETOOTH-15107-kw45-check-and-fix-misra-for-all-digital_key-and-localization-projects to develop/main

Squashed commit of the following:

commit 942d93f5dc195c46d06115c04ababf0b65ba96a8
Author: Andreea Binca <andreea.binca@nxp.com>
Date:   Wed Mar 26 12:01:45 2025 +0200

    [BLUETOOTH-15107][KW45] Check and fix MISRA for all digital_key and localization projects

+++ examples/digital_key_car_anchor/app_digital_key_car_anchor.c
@@ -543,7 +543,7 @@ void APP_UserInterfaceEventHandler(void *pData)
                 default:
                 {
                     shell_write("\r\nHandover error ");
-                    shell_writeDec(pEventData->eventData.handoverError);
+                    shell_writeDec((uint32_t)(pEventData->eventData.handoverError));
                     shell_write(".\r\n");
                     shell_cmd_finished();
                 }
